backend/requirements.txt
# Manga Translator Backend Dependencies

# FastAPI Framework
fastapi==0.121.0
uvicorn==0.38.0

# Data validation
pydantic==2.12.4

# HTTP
requests==2.32.5

# Config
python-dotenv==1.2.1

# Image processing
pillow==12.0.0

# AI Translation (OpenAI-compatible)
openai==2.7.1

# OCR (optional - enable via PADDLE_OCR_ENABLED=true)
# paddlepaddle==3.2.1
# paddleocr==3.3.1
# opencv-contrib-python==4.10.0.84
# numpy==2.2.6
# pyclipper==1.3.0.post6
# shapely==2.1.2
